+++ Board-code/HiL_Board/HiL_code/Src/HiL_MSGQ.c
@@ -1,19 +1,19 @@
 /*
- * HiL_Init_MSGQ.c
+ * HiL_MSGQ.c
  *
- *  Created on: Nov 25, 2022
- *      Author: Olle
+ *  Created on: Dec 12, 2022
+ *      Author: OlleBlo
  *
- *      This file contains the initialization for the message queues in use in freeRTOS
- *
+ *      This file contains the initialization for the message queues and other message queue related functions in use in freeRTOS
  */
 
-#include "HiL_Init_MSGQ.h"
+
+#include "HiL_MSGQ.h"
 
 extern osMessageQueueId_t USB_MSGQ_Rx;
 //extern osMessageQueueId_t USB_MSGQ_Tx;
 
-int HiL_Init_MSGQ (void) {
+uint8_t HiL_Init_MSGQ (void) {
 
 	USB_MSGQ_Rx = osMessageQueueNew(HiL_USB_MSQG_len, sizeof(MSGQ_obj), NULL);
   if (USB_MSGQ_Rx == NULL) {
